+++ stencil-workspace/src/components/modus-button/modus-button.tsx
@@ -1,5 +1,14 @@
 // eslint-disable-next-line
-import { Component, Prop, h, Event, EventEmitter, Element, State, Listen } from '@stencil/core';
+import {
+  Component,
+  Prop,
+  h,
+  Event,
+  EventEmitter,
+  Element,
+  State,
+  Listen,
+} from '@stencil/core';
 
 @Component({
   tag: 'modus-button',
@@ -58,13 +67,17 @@ export class ModusButton {
   }
 
   render(): unknown {
-    const className = `${this.classBySize.get(this.size)} ${this.classByColor.get(this.color)} ${this.classByButtonStyle.get(this.buttonStyle)}`;
+    const className = `${this.classBySize.get(
+      this.size
+    )} ${this.classByColor.get(this.color)} ${this.classByButtonStyle.get(
+      this.buttonStyle
+    )}`;
 
     return (
       <button
-        aria-disabled={this.disabled}
+        aria-disabled={this.disabled ? 'true' : undefined}
         aria-label={this.ariaLabel}
-        aria-pressed={this.pressed}
+        aria-pressed={this.pressed ? 'true' : undefined}
         class={className}
         disabled={this.disabled}
         onClick={() => (!this.disabled ? this.buttonClick.emit() : null)}