+++ CHANGELOG.md
@@ -6,6 +6,8 @@ View all releases at: https://github.com/trimble-oss/modus-web-components/releas
 
 ## Unreleased
 
+## 0.1.18 - 2022-12-09
+
 ### Added
 
 - Added props `draggableItem`, `droppableItem`, `editable` and `label` slot in `modus-tree-view-item` component